requirements.txt
fastapi==0.115.0
uvicorn==0.27.1
pydantic[email]==2.8.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
openai==1.3.7
elevenlabs==0.2.27
google-cloud-firestore==2.16.1
google-cloud-tasks==2.13.2
google-cloud-secret-manager==2.16.2
google-cloud-storage==2.14.0
google-cloud-resource-manager==1.14.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]>=3.4.0
httpx==0.26.0
redis==6.2.0
PyYAML==6.0.1
watchdog==4.0.0
textstat==0.7.3
language_tool_python==2.7.1
python-json-logger==2.0.7
hiredis==2.3.2
prometheus_client
python-multipart
tenacity

# Core test dependencies for local/dev use
pytest
pytest-mock
pytest-cov
pytest-asyncio
pytest-xdist
